`@remotion/serverless`: Extract single-function renderer
